Rewrite the XML parsing components.
The parsing code is now isolated, and theoretically other markup languages or
file formats could be used.
There are two out standing problems with this code because I am working on a
Windows computer and have no C compiler or symlink capacity. There may therefore
be uncaught syntax errors, and there should be a format.h symlink to
xml-format.h.